pipeline/pyproject.toml
[project]
name = "quality-share-pipeline"
version = "0.1.0"
description = "AI Curation Pipeline for Quality Share"
authors = [
    {name = "thakurajayL",email = "thakur.ajay@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "feedparser (>=6.0.12,<7.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.14.2,<5.0.0)",
    "trafilatura (>=2.0.0,<3.0.0)",
    "langchain (>=1.0.0,<2.0.0)",
    "openai (>=2.5.0,<3.0.0)",
    "langchain-openai (>=1.0.0,<2.0.0)",
    "pygithub (>=2.8.1,<3.0.0)",
    "python-frontmatter (>=1.0.0,<2.0.0)",
    "PyYAML (>=6.0.1,<7.0.0)"
]

[[tool.poetry.packages]]
include = "src"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "ruff (>=0.14.1,<0.15.0)"
]

[tool.pytest.ini_options]
pythonpath = ["src"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
